requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
torch==2.2.0
transformers==4.34.1
sounddevice==0.4.6
numpy==1.24.3
sentence-transformers==2.2.2
python-dotenv==1.0.0